requirements.txt
# TNSIM requirements for pip
# # Usage: pip install -r requirements.txt

pyserial
pylint
ConfigArgParse
PySide2
py-cui

# For documentation:
Sphinx
sphinx_rtd_theme
recommonmark